Split up e4x custom-visitors and remove e4x from emitter.js
